requirements.txt
aiofiles==24.1.0
alabaster==0.7.16
annotated-types==0.7.0
anyio==4.4.0
astroid==2.15.8
attrs==24.1.0
Authlib==1.3.1
Babel==2.15.0
bandit==1.7.9
black==24.8.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.6.0
cryptography==43.0.0
Deprecated==1.2.14
dill==0.3.8
distlib==0.3.8
distro==1.9.0
docstring_parser==0.16
docutils==0.20.1
dodgy==0.2.1
dparse==0.6.4b0
exceptiongroup==1.2.2
filelock==3.12.2
flake8==5.0.4
flake8-polyfill==1.0.2
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-aiplatform==1.60.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpc-google-iam-v1==0.13.1
grpcio==1.64.0
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
identify==2.6.0
idna==3.7
imagesize==1.4.1
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.4
jiter==0.5.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lazy-object-proxy==1.10.0
mando==0.7.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mccabe==0.7.0
mdurl==0.1.2
mypy==1.11.1
mypy-extensions==1.0.0
nodeenv==1.9.1
numpy==2.0.1
objprint==0.2.3
openai==1.40.2
packaging==24.1
pathspec==0.12.1
pbr==6.0.0
pep8-naming==0.10.0
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.8.0
prospector==1.10.3
proto-plus==1.24.0
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.9.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeps==1.12.20
pydocstyle==6.3.0
pyflakes==2.5.0
Pygments==2.18.0
pylint==2.17.7
pylint-celery==0.3
pylint-django==2.5.3
pylint-flask==0.6
pylint-plugin-utils==0.7
pyright==1.1.374
pytest==8.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
radon==6.0.1
referencing==0.35.1
requests==2.32.3
requirements-detector==1.2.2
rich==13.7.1
rpds-py==0.19.1
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safety==3.2.4
safety-schemas==0.0.2
schema==0.7.7
semver==3.0.2
setoptconf-tmp==0.3.1
shapely==2.0.5
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
snowballstemmer==2.2.0
Sphinx==7.4.7
sphinx-markdown-tables==0.0.17
sphinx-rtd-theme==2.0.0
sphinx-tabs==3.4.5
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-mermaid==0.9.2
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stdlib-list==0.10.0
stevedore==5.2.0
tenacity==9.0.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.13.0
tqdm==4.66.5
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.2
vertexai==1.60.0
virtualenv==20.26.3
vulture==2.11
wrapt==1.16.0
